day-11: part II - solved

+++ src/main/java/pl/szymhu/day11/MonkeyThrowResult.java
@@ -1,4 +1,4 @@
 package pl.szymhu.day11;
 
-record MonkeyThrowResult(int destination, int item) {
+record MonkeyThrowResult(int destination, long item) {
 }
